Make http/s submodules single export
  - Isolate helper functions into siloed utils folders
    for full unit testing.
  - Modify tests to account for new structure.
