+++ lib/transports/http.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Google Inc. All Rights Reserved.
+ * Copyright 2017 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@
 var url = require('./http/url');
 var headers = require('./http/headers');
 var payload = require('./http/payload');
-var utils = require('./http/utils');
+var utils = require('./http/utils/http.js');
 
 function formulateRequest (generatedOptions, cb) {
   var body = '';
@@ -45,9 +45,9 @@ function formulateRequest (generatedOptions, cb) {
 
 function request (givenOptions, cb) {
   return formulateRequest(
-    url.generateUrl(
-      headers.generateHeaders(
-        payload.generateRequestPayload({}, givenOptions),
+    url(
+      headers(
+        payload({}, givenOptions),
         givenOptions
       ),
       givenOptions
