+++ lib/transports/http/headers.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Google Inc. All Rights Reserved.
+ * Copyright 2017 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,86 +15,32 @@
  */
 
 'use strict';
-var merge = require('lodash.merge');
-var isString = require('lodash.isstring');
-var pkg = require('../../../package.json');
 
-var USER_AGENT = 'google-api-nodejs-client/'+pkg.version;
-var CONTENT_TYPES = {
-  json: 'application/json',
-  form: 'application/x-www-form-urlencoded'
-};
-var BODY_REQUESTS = ['POST', 'PUT', 'PATCH'];
+var utils = require('./utils/headers.js');
 
-function extractGivenHeaders (generatedOptions, givenOptions) {
-  if (givenOptions.headers) {
-    return merge({}, generatedOptions, {
-      requestOptions: {
-         headers: givenOptions.headers
-      }
-    });
-  }
-  return generatedOptions;
-}
-
-function addUserAgentHeaders (generatedOptions) {
-  return merge({}, generatedOptions, {
-    requestOptions: {
-      headers: {
-        'User-Agent': USER_AGENT
-      }
-    }
-  });
-}
-
-function addRequestMethod (generatedOptions, givenOptions) {
-  return merge({}, generatedOptions, {
-    requestOptions: {
-      method: isString(givenOptions.method) ?
-        givenOptions.method.toUpperCase() : 'GET'
-    }
-  });
-}
-
-function addPostMethodHeaders (generatedOptions, givenOptions) {
-  var hasBody = BODY_REQUESTS
-    .indexOf(generatedOptions.requestOptions.method) > -1;
-  if (hasBody === true) {
-    return merge({}, generatedOptions, {
-      hasBody: hasBody,
-      requestOptions: {
-        headers: {
-          'Content-Type': generatedOptions.requestPayload.isJSON ?
-            CONTENT_TYPES.json : CONTENT_TYPES.form,
-          'Content-Length': Buffer.byteLength(
-            generatedOptions.requestPayload.payload)
-        }
-      }
-    });
-  } else if (!hasBody && givenOptions && (givenOptions.json === true)) {
-    // TODO: add an explicit unit test for this branch
-    return merge({}, generatedOptions, {
-      requestOptions: {
-        headers: {
-          'Content-Type': CONTENT_TYPES.json
-        }
-      }
-    });
-  }
-  return merge({}, generatedOptions, {hasBody: hasBody});
-}
-
+/**
+ * Invokes the headers constituent functions in correct order to augment the
+ * generatedOptions with header information about the request.
+ * @function generateHeaders
+ * @param {Object} generatedOptions - the options used to execute the request
+ * @param {Object} givenOptions - the user-provided request options
+ * @return {Object} - returns a new object which is an amalgam of originally
+ *  given generatedOptions and any necessary header information.
+ */
 function generateHeaders (generatedOptions, givenOptions) {
-  return addPostMethodHeaders(addRequestMethod(addUserAgentHeaders(
-    extractGivenHeaders(generatedOptions, givenOptions)), givenOptions),
-    givenOptions);
+  return utils.addPostMethodHeaders(
+    utils.addRequestMethod(
+      utils.addUserAgentHeaders(
+        utils.extractGivenHeaders(generatedOptions, givenOptions)
+      ),
+      givenOptions
+    ),
+    givenOptions
+  );
 }
 
-module.exports = {
-  extractGivenHeaders: extractGivenHeaders,
-  addUserAgentHeaders: addUserAgentHeaders,
-  addRequestMethod: addRequestMethod,
-  addPostMethodHeaders: addPostMethodHeaders,
-  generateHeaders: generateHeaders,
-  USER_AGENT: USER_AGENT
-};
+// Set the readonly method for user agent as publically exposed for testing
+// purposes
+generateHeaders.USER_AGENT = utils.USER_AGENT;
+
+module.exports = generateHeaders;
