+++ lib/transports/http/payload.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Google Inc. All Rights Reserved.
+ * Copyright 2017 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,29 +16,9 @@
 
 'use strict';
 var merge = require('lodash.merge');
-var isString = require('lodash.isstring');
-var isObject = require('lodash.isobject');
 var has = require('lodash.has');
 var queryString = require('querystring');
-
-function extractFormData (formData) {
-  return queryString.stringify(formData);
-}
-
-function extractJsonData (jsonProp, bodyProp, formProp) {
-  if (isString(jsonProp)) {
-    return jsonProp;
-  } else if (isObject(jsonProp)) {
-    return JSON.stringify(jsonProp);
-  } else if (jsonProp === true) {
-    if (isObject(bodyProp)) {
-      return JSON.stringify(bodyProp);
-    } else if (isObject(formProp)) {
-      return JSON.stringify(formProp);
-    }
-  }
-  return '';
-}
+var extractJsonData = require('./utils/payload.js').extractJsonData;
 
 function generateRequestPayload (generatedOptions, givenOptions) {
   var out = generatedOptions;
@@ -62,7 +42,7 @@ function generateRequestPayload (generatedOptions, givenOptions) {
     out = merge({}, generatedOptions, {
       requestPayload: {
         isJSON: false,
-        payload: extractFormData(givenOptions.form)
+        payload: queryString.stringify(givenOptions.form)
       }
     });
   }
@@ -76,8 +56,4 @@ function generateRequestPayload (generatedOptions, givenOptions) {
   return out;
 }
 
-module.exports = {
-  extractFormData: extractFormData,
-  extractJsonData: extractJsonData,
-  generateRequestPayload: generateRequestPayload
-};
+module.exports = generateRequestPayload;
