+++ lib/transports/http/utils/http.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Google Inc. All Rights Reserved.
+ * Copyright 2017 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,6 +22,40 @@ var isString = require('lodash.isstring');
 var isObject = require('lodash.isobject');
 var noOp = require('lodash.noop');
 var has = require('lodash.has');
+
+function getTransport (isUsingHttps) {
+  if (isUsingHttps) {
+    return https;
+  }
+  return http;
+}
+
+
+function attemptBodyParse (body) {
+  try {
+    return JSON.parse(body);
+  } catch (e) {
+    return body;
+  }
+}
+
+function wrapGivenCallback (cb) {
+  if (!isFunction(cb)) {
+    return noOp;
+  }
+  return function (err, res, body) {
+    var parsedBody, parsedError;
+    if (err || !body) {
+      return cb(err, body, res);
+    }
+    parsedBody = attemptBodyParse(body);
+    parsedError = generateErrorInformation(parsedBody, res);
+    if (parsedError) {
+      parsedBody = null;
+    }
+    cb(parsedError, parsedBody, res);
+  };
+}
 
 function generateErrorInformation (body, res) {
   var err = null;
@@ -48,40 +82,6 @@ function generateErrorInformation (body, res) {
   }
   return err;
 }
-
-function wrapGivenCallback (cb) {
-  if (!isFunction(cb)) {
-    return noOp;
-  }
-  return function (err, res, body) {
-    var parsedBody, parsedError;
-    if (err || !body) {
-      return cb(err, body, res);
-    }
-    parsedBody = attemptBodyParse(body);
-    parsedError = generateErrorInformation(parsedBody, res);
-    if (parsedError) {
-      parsedBody = null;
-    }
-    cb(parsedError, parsedBody, res);
-  };
-}
-
-function attemptBodyParse (body) {
-  try {
-    return JSON.parse(body);
-  } catch (e) {
-    return body;
-  }
-}
-
-function getTransport (isUsingHttps) {
-  if (isUsingHttps) {
-    return https;
-  }
-  return http;
-}
-
 
 module.exports = {
   wrapGivenCallback: wrapGivenCallback,
