+++ lib/transports/http/utils/payload.js
@@ -15,16 +15,24 @@
  */
 
 'use strict';
-var utils = require('./utils/url.js');
+var isString = require('lodash.isstring');
+var isObject = require('lodash.isobject');
 
-function generateUrl (generatedOptions, givenOptions) {
-  return utils.extractUrlComponents(
-    utils.isUsingHttps(
-      utils.createUrlObject(
-        utils.extractRawHref(generatedOptions, givenOptions)
-      )
-    )
-  );
+function extractJsonData (jsonProp, bodyProp, formProp) {
+  if (isString(jsonProp)) {
+    return jsonProp;
+  } else if (isObject(jsonProp)) {
+    return JSON.stringify(jsonProp);
+  } else if (jsonProp === true) {
+    if (isObject(bodyProp)) {
+      return JSON.stringify(bodyProp);
+    } else if (isObject(formProp)) {
+      return JSON.stringify(formProp);
+    }
+  }
+  return '';
 }
 
-module.exports = generateUrl;
+module.exports = {
+  extractJsonData: extractJsonData
+};
